move weaviate writer to connectors/weaviate/. also fix an IT
