+++ lib/sycamore/sycamore/connectors/weaviate/__init__.py
@@ -1,0 +1,13 @@
+from sycamore.connectors.weaviate.weaviate_writer import (
+    WeaviateDocumentWriter,
+    WeaviateCrossReferenceWriter,
+    WeaviateClientParams,
+    WeaviateTargetParams,
+)
+
+__all__ = [
+    "WeaviateDocumentWriter",
+    "WeaviateCrossReferenceWriter",
+    "WeaviateClientParams",
+    "WeaviateTargetParams",
+]
