+++ lib/sycamore/sycamore/connectors/weaviate/weaviate_writer.py
